+++ src/lib.rs
@@ -7,6 +7,9 @@ pub mod engine;
 pub mod diode;
 pub mod isine;
 
+pub mod capacitor;
+
 pub mod wavewriter;
+