+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [UNRELEASED] Capacitors
+Deal with capacitors
+
+### Added
+- Capacitor model for DC operating point analysis
+
+
 
 ## [0.5.0] Transient Analysis
 Added iteration count based transient analysis.